moving implementations to longhands
